Reformat, move file service_resolvable.rs
