+++ aries_vcx/src/did_doc/model.rs
@@ -1,12 +1,4 @@
-
 use serde::ser::{Serialize, Serializer};
-
-
-
-
-
-
-
 
 pub const CONTEXT: &str = "https://w3id.org/did/v1";
 pub const KEY_TYPE: &str = "Ed25519VerificationKey2018";
@@ -15,11 +7,11 @@ pub const KEY_AUTHENTICATION_TYPE: &str = "Ed25519SignatureAuthentication2018";
 #[derive(Debug, Deserialize, Serialize, Clone, PartialEq)]
 pub struct Ed25519PublicKey {
     pub id: String,
+    // all list of types: https://w3c-ccg.github.io/ld-cryptosuite-registry/
     #[serde(rename = "type")]
     pub type_: String,
-    // all list of types: https://w3c-ccg.github.io/ld-cryptosuite-registry/
     pub controller: String,
-    #[serde(rename = "publicKeyBase58")] // todo: rename to publicKeyMultibase
+    #[serde(rename = "publicKeyBase58")]
     pub public_key_base_58: String,
 }
