+++ aries_vcx/src/handlers/mod.rs
@@ -10,7 +10,6 @@ pub mod connection;
 pub mod issuance;
 pub mod proof_presentation;
 pub mod out_of_band;
-pub mod service_resolvable;
 
 impl From<ConnectionState> for u32 {
     fn from(state: ConnectionState) -> u32 {
